% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{append_caption_docx}
\alias{append_caption_docx}
\title{Append Caption to Word Document}
\usage{
append_caption_docx(word_doc, caption)
}
\arguments{
\item{word_doc}{The Word document object to which the caption will be
appended.}

\item{caption}{A character vector containing the caption text to be appended
to the Word document. If NULL or not provided, no caption will
be appended.}
}
\value{
The modified Word document object with the caption appended (if
provided) or the original Word document object if no caption is
provided.
}
\description{
This function appends a caption to a Word document. It is designed to be used
internally by docx-specific export functions (e.g., \code{ggplot_to_docx})
}
\examples{

library(outbox)

# Example of internal call that might be made by ggplot_to_docx()
word_doc <- officer::read_docx()
caption <- 'This is a sample caption.'
append_caption_docx(word_doc, caption)

}
