% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{append_caption_xlsx}
\alias{append_caption_xlsx}
\title{Append Caption to XLSX Sheet}
\usage{
append_caption_xlsx(output_wb, sheet_name, caption)
}
\arguments{
\item{output_wb}{The output workbook object to which the caption will be
appended.}

\item{sheet_name}{A character string representing the name of the sheet to
which the caption will be appended.}

\item{caption}{A character string containing the caption text to be appended
to the sheet. If NULL or not provided, no caption will be
appended.}
}
\value{
The modified output workbook object with the caption appended (if
provided) or the original output workbook object if no caption is
provided.
}
\description{
This function appends a caption to an XLSX sheet. It is designed to be used
internally by xlsx-specific export functions (e.g., \code{ggplot_to_xlsx})
}
\examples{
# Example of internal call that might be made by ggplot_to_xlsx()
sheet_name <- 'Sheet1'
caption <- 'This is a sample caption.'

# avoid using pipes with openxlsx
output_wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(output_wb, sheetName = sheet_name)

output_wb <- outbox:::append_caption_xlsx(output_wb, sheet_name, caption)

}
