% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-naming.R
\name{append_date}
\alias{append_date}
\title{Append today's date to the end of a file path}
\usage{
append_date(path)
}
\arguments{
\item{path}{The file path string (including file name and extension) to which the
current date should be appended.}
}
\value{
A modified string filepath with the date added (just before the
file extension at the end).
}
\description{
Append today's date to the end of a file path
}
\examples{

my_path <- 'my_dir/my_file.xlsx'
append_date(my_path)
}
