% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{construct_output_function}
\alias{construct_output_function}
\title{Construct Output Function Name}
\usage{
construct_output_function(x, path)
}
\arguments{
\item{x}{An object from which to detect the output type.}

\item{path}{A character vector representing the file path from which to detect
the output extension.}
}
\value{
A character vector representing the name of the constructed output
function. If the output type or extension is not supported, the function
will throw an error.
}
\description{
This function is used by the generic \code{write_output()} function to make
the appropriate specific output function call (e.g.,
\code{gtsummary_to_docx}) under the hood.
}
\details{
It constructs the name of the specific output function to be used (e.g.,
\code{gtsummary_to_docx}), based on the detected output type and output
extension. The output function name is in the format
"{output_type}\emph{to}{output_ext}".

This function is intended for internal use only. It is not exported to the
end user.
}
\examples{
my_ggplot <- ggplot2::ggplot(mtcars, ggplot2::aes(x = mpg, y = hp))
my_path <- "output.xlsx"
construct_output_function(my_ggplot, my_path)
}
