% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{create_docx}
\alias{create_docx}
\title{Create an empty Word Document (docx)}
\usage{
create_docx(path, toc = TRUE)
}
\arguments{
\item{path}{The file path where the Word Document (docx) will be created.}

\item{toc}{Logical, indicating whether to include a Table of Contents (TOC) in the
created document. The default is TRUE.}
}
\value{
An invisible Word Document (docx) object representing the empty Word Document.
}
\description{
This function creates an empty Word Document (docx) at the specified path. It is designed
to be used internally for initializing docx files before adding content or sections.
}
\details{
This function creates an empty Word Document (docx) and returns the document
object, which can be used to add content, paragraphs, sections, or Table of Contents (TOC).
By default, a Table of Contents (TOC) is included in the document.
}
\examples{

\dontrun{
# Create an empty Word Document (docx) named 'example.docx'
outbox:::create_docx('example.docx')

# Create an empty Word Document (docx) with Table of Contents (TOC) disabled
outbox:::create_docx('example_without_toc.docx', toc = FALSE)
}
}
\seealso{
Other docx-related functions: 
\code{\link{append_caption_docx}()},
\code{\link{write_output}()}
}
\concept{docx-related functions}
\keyword{internal}
