% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{create_xlsx}
\alias{create_xlsx}
\title{Create an empty XLSX file}
\usage{
create_xlsx(path)
}
\arguments{
\item{path}{The file path where the XLSX file will be created.}
}
\value{
An invisible workbook object representing the empty XLSX file.
}
\description{
This function creates an empty XLSX file at the specified path. It is designed
to be used internally for initializing XLSX files before adding data or sheets.
}
\details{
This function creates an empty XLSX file and returns the workbook
object, which can be used to add data or sheets to the file.
}
\examples{
# Create an empty XLSX file named 'example.xlsx'
create_xlsx('example.xlsx')

}
\seealso{
Other xlsx-related functions: 
\code{\link{append_caption_xlsx}()},
\code{\link{write_output}()}
}
\concept{xlsx-related functions}
\keyword{internal}
