% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{detect_output_ext}
\alias{detect_output_ext}
\title{Detect the Output Extension from a File Path}
\usage{
detect_output_ext(path)
}
\arguments{
\item{path}{A character vector representing the file path from which to
detect the output extension.}
}
\value{
A character vector indicating the detected output extension. If the
detected extension is not one of the supported extensions ('xlsx' or 'docx'),
the function will throw an error.
}
\description{
This function detects the output extension from a given file path. It checks
if the file extension matches one of the supported extensions ('xlsx' or
'docx'), and throws an error if the extension is not a supported type.
}
\details{
This function is used internally by \code{write_output()}. Generally, the end
user should not have a reason to call it directly. It is exported instead as
a convenience function, in case the user needs to debug something
(e.g., to understand how the \code{write_output()} function is assessing
their output objects internally).
}
\examples{
library(outbox)

path <- 'example.xlsx'
detect_output_ext(path)

}
\seealso{
Other write_output-related functions: 
\code{\link{construct_output_function}()},
\code{\link{detect_output_type}()},
\code{\link{last_path}()},
\code{\link{write_output}()}
}
\concept{write_output-related functions}
