% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{detect_output_type}
\alias{detect_output_type}
\title{Detect the Output Type from an Object}
\usage{
detect_output_type(x)
}
\arguments{
\item{x}{An output object (e.g. table) from which to detect the output type.}
}
\value{
A character vector indicating the detected output type. If the
detected class is not one of the supported classes ('gtsummary' or 'ggplot'),
the function will throw an error.
}
\description{
This function detects the output type from a given object. It checks if the
object's class matches one of the supported classes ('gtsummary' or
'ggplot'), and throws an error if the extension is not a supported type.
}
\details{
This function is used internally by \code{write_output()}. Generally, the end
user should not have a reason to call it directly. It is exported instead as
a convenience function, in case the user needs to debug something
(e.g., to understand how the \code{write_output()} function is assessing
their output objects internally).
}
\examples{
library(outbox)

my_ggplot <- ggplot2::ggplot(mtcars, ggplot2::aes(x = mpg, y = hp))
detect_output_type(my_ggplot)

}
\seealso{
Other write_output-related functions: 
\code{\link{construct_output_function}()},
\code{\link{detect_output_ext}()},
\code{\link{last_path}()},
\code{\link{write_output}()}
}
\concept{write_output-related functions}
