% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{supported}
\alias{supported}
\title{Get Supported Output Object Classes and File Types}
\usage{
supported(what = NULL)
}
\arguments{
\item{what}{A character string specifying what to retrieve. Possible values
are "classes" and "extensions". If \code{what} is \code{NULL}, the function
will return a list containing both supported classes and extensions.}
}
\value{
If \code{what} is \code{NULL}, the function returns a list containing
two character vectors: \code{classes} and \code{extensions}, representing
the supported output object classes and file extensions, respectively. If
\code{what} is specified, the function will return the corresponding vector
of supported classes or extensions.
}
\description{
This function returns the list of supported output object classes and file
types by the outbox R package. It can be used to query the supported classes
and extensions for the \code{write_output()} function.
}
\examples{
# Get the list of supported output classes and extensions
supported()

# Get the supported output object classes
supported('classes')

# Get the supported file extensions
supported('extensions')

}
