% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-naming.R
\name{append_date}
\alias{append_date}
\title{Append today's date to the end of a file path}
\usage{
append_date(path)
}
\arguments{
\item{path}{The file path string (including file name and extension) to which
the current date should be appended.}
}
\value{
A modified string filepath with the date added (just before the
file extension at the end).
}
\description{
Append today's date to the end of a file path
}
\details{
When code is run multiple times, it is sometimes desireable to maintain its
previous output files, rather than overwriting them (e.g., to compare old vs.
new in the case of a bug). One way to facilitate that is to append the
current date to your outbox file path at the top of a document. That way,
if the code is run on a new day, nothing will be overwritten from the
previous day.
}
\examples{

my_path <- 'my_dir/my_file.xlsx'
append_date(my_path)
}
