% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{last_path}
\alias{last_path}
\title{Retrieve the last path used in write_output()}
\usage{
last_path()
}
\value{
The last path used in \code{write_output()}, or \code{NULL} if no
path has been used yet.
}
\description{
This function returns the last path used in a previous call to
write_output(). If no path has been used yet, it returns NULL. It is desiged
to be used in conjunction with \code{write_output()}.
}
